Moverio - Single Point Active Alignment Method
This code facilitates the conduction of a Single Point Active Alignment
Method calibration on the Epson Moverio BT-200.

The calibration may be performed for the Left or the Right eye
individually (stereo calibration is currently not supported).

The results of the calibration are placed in the 'External Storage' of
the device in the 'Download' directory under a folder called
'SPAAM_Calib' ('SDCard'/Download/SPAAM_Calib). The files themselves are
'Right.calib' and 'Left.calib' respectively for each eye result.

After the eye to calibrate is selected, the previously recorded result
(if available) will be loaded for the user to see. This result, however,
will be completely replaced after the new calibration is performed.

See the Readme.txt file for more information.