Required Libraries and Usage Instructions
ThirdPartyLibraries.zip contains all of the additional jar files
required (except for the JRE rt.java which you will have to link to on
your local machine). The libraries included with this file are:

The Vuforia library used to build this project
The Jama math library used to perform the SVD calculations for SPAAM
The Moverio BT-200 interface libraries for setting Moverio specific
properties at runtime.

The Readme.txt file contains all of the instructions required to setup,
build, and use the application.